ragagent/requirements.txt

# Core LLM & LangChain
google-generativeai>=0.3.0
langchain>=0.1.0
langchain-community>=0.0.29
sentence-transformers>=2.2.2
langchain-google-genai


# Vector DBs
cassio>=0.1.4
faiss-cpu>=1.7.2
cassandra-driver>=3.28.0
astrapy
langchain-astradb
tiktoken



# FastAPI + Backend
fastapi>=0.68.0
uvicorn>=0.15.0
python-dotenv>=0.19.0
passlib[bcrypt]>=1.7.4
python-jose>=3.3.0
PyJWT>=2.0.0


# PDF, OCR, NLP
pypdf>=3.0.0
pdfplumber
unstructured>=0.7.0
opencv-python-headless
pytesseract
numpy>=1.21.0
pandas>=1.3.0
flashrank>=0.1.0

pydantic-settings>=2.0.0
pydantic>=2.0.0
python-multipart
bcrypt==3.2.2
passlib[bcrypt]==1.7.4

streamlit
requests
